Move `createSchemaField` into Client.prototype, because it follows the same pattern as all the other methods, even though it's only used in testing.
