+++ lib/solr.js
@@ -884,4 +884,60 @@ Client.prototype.ping = function (callback) {
   return this.get(this.ADMIN_PING_HANDLER, callback);
 };
 
+Client.prototype.createSchemaField = function (fieldName, fieldType, cb) {
+  const json = JSON.stringify({
+    'add-field': {
+      name: fieldName,
+      type: fieldType,
+      multiValued: false,
+      stored: true,
+    },
+  });
+
+  const callback = (err, data) => {
+    if (err) {
+      // TODO: We should handle this in a more robust way in the future, but
+      //   there is a difference between default setup in Solr 5 and Solr 8, so
+      //   some fields already exist in Solr 8. Hence if that's the case, we
+      //   just ignore that.
+      console.warn(err.message);
+    }
+    cb(undefined, data);
+  };
+
+  const requestOptions = {
+    host: this.options.host,
+    port: this.options.port,
+    method: 'POST',
+    headers: {
+      'content-type': 'application/json; charset=utf-8',
+      'content-length': Buffer.byteLength(json),
+      accept: 'application/json; charset=utf-8',
+    },
+    path: 'http://localhost:8983/solr/testcore/schema',
+    family: this.options.ipVersion,
+  };
+
+  if (this.options.authorization) {
+    requestOptions.headers.authorization = this.options.authorization;
+  }
+
+  if (this.options.agent) {
+    requestOptions.agent = this.options.agent;
+  }
+
+  const request = pickProtocol(this.options.secure).request(requestOptions);
+
+  request.on('response', handleJSONResponse(request, this.options.bigint, callback));
+
+  request.on('error', function onError(err) {
+    if (callback) callback(err, null);
+  });
+
+  request.write(json);
+  request.end();
+
+  return request;
+}
+
 Promise.promisifyAll(Client.prototype);
