+++ test/core-query-test.js
@@ -14,13 +14,12 @@ const figc = require('figc'),
 const config = figc(__dirname + '/config.json');
 const client = solr.createClient(config.client);
 [config.client.path, config.client.core].join('/').replace(/\/$/, '');
-const { createSchemaField } = require('./utils/schemaHelper');
 
 describe('Client#createQuery', function () {
   before((cb) => {
-    createSchemaField(client, 'title', 'text_general', () => {
-      createSchemaField(client, 'name', 'text_general', () => {
-        createSchemaField(client, 'author', 'text_general', cb);
+    client.createSchemaField('title', 'text_general', () => {
+      client.createSchemaField('name', 'text_general', () => {
+        client.createSchemaField('author', 'text_general', cb);
       });
     });
   });
